split consts and utils by dirs
