+++ src/consts/render.js
@@ -1,0 +1,7 @@
+export const RenderPosition = {
+  AFTERBEGIN: 'afterbegin',
+  BEFOREEND: 'beforeend',
+  BEFOREBEGIN: 'beforebegin',
+  AFTEREND: 'afterend',
+  REPLACEWITH: 'replacewith',
+};
