+++ src/mock/cities.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils';
+import { getRandomInteger } from '../utils/common';
 
 const names = ['Amsterdam', 'Chamonix', 'Geneva', 'Moscow', 'Los-Angeles'];
 const description =
