+++ src/mock/const.js
@@ -73,10 +73,3 @@ export const offersList = [
     type: 'restaurant',
   },
 ];
-
-export const RenderPosition = {
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  BEFOREBEGIN: 'beforebegin',
-  AFTEREND: 'afterend',
-};
