+++ src/view/message.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils';
+import { createElement } from '../utils/render';
 
 const getTemplate = (message) => {
   return `<p class="trip-events__msg">${message}</p>`;
