+++ src/view/sorting.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils';
+import { createElement } from '../utils/render';
 
 const getTemplate = () => {
   return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
